Restructured architecture
